refactor: 임시 토큰 리펙토링
클래스명 변경
코드 개행
